+++ src/Illuminate/Bus/BusServiceProvider.php
@@ -87,7 +87,7 @@ class BusServiceProvider extends ServiceProvider implements DeferrableProvider
             );
         });
 
-        $this->app->bind(ResumeStateRepository::class,CacheResumeStateRepository::class);
+        $this->app->bind(ResumeStateRepository::class, CacheResumeStateRepository::class);
     }
 
     /**
@@ -103,6 +103,7 @@ class BusServiceProvider extends ServiceProvider implements DeferrableProvider
             QueueingDispatcherContract::class,
             BatchRepository::class,
             DatabaseBatchRepository::class,
+            ResumeStateRepository::class,
         ];
     }
 }
