+++ src/Illuminate/Workflow/WorkflowOrchestrator.php
@@ -2,6 +2,8 @@
 
 namespace Illuminate\Workflow;
 
+use Illuminate\Bus\Workflow\Workflow;
+
 class WorkflowOrchestrator {
     public function __construct(protected Workflow $workflow)
     {
